refactor code structure for service/nft
