+++ service/nft/keys.go
@@ -1,8 +1,8 @@
-package service
+package nft
 
 const (
 	PROTOCOL            = "https"
 	DOMAIN              = "deep-index.moralis.io"
 	MORALIS_API_VERSION = "api/v2"
-	MORALIS_API = PROTOCOL + "://" + DOMAIN + "/" + MORALIS_API_VERSION + "/"
+	MORALIS_API         = PROTOCOL + "://" + DOMAIN + "/" + MORALIS_API_VERSION + "/"
 )
